requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
mysql-connector-python==8.2.0
pymysql==1.1.0
cryptography==41.0.7 